Component-ized the app
